Fix loading ais/amg state from embeddings in 3d annotator
